pyproject.toml
[project]
name = "cv-mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "docx2txt>=0.9",
    "google-api-python-client>=2.188.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "httpx>=0.28.1",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.7",
    "lxml>=6.0.2",
    "mcp>=1.26.0",
    "openai>=2.16.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
]
